requirements.txt
attrs==20.3.0
autopep8==1.5.4
base58==2.0.1
bitarray==1.2.2
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
coinpaprika==0.1.0
cytoolz==0.11.0
dnspython==2.0.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-translate==3.0.1
googleapis-common-protos==1.52.0
grpcio==1.33.2
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
jsonschema==3.2.0
langdetect==1.0.8
libcst==0.3.14
lru-dict==1.1.6
multiaddr==0.0.9
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycoingecko==1.4.0
pycryptodome==3.9.9
pymongo==3.11.1
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rlp==2.0.1
rsa==4.6
six==1.15.0
toml==0.10.2
toolz==0.11.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
varint==1.0.2
web3==5.13.0
websocket-client==0.57.0
websockets==8.1
